# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ENUMERA API",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "name": "ENUMERA",
            "url": "https://enumera.tech",
            "email": "enumera@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Autentification"
                ],
                "summary": "Création d'un compte utilisateur",
                "description": "Création d'un nouveau compte utilisateur",
                "operationId": "receiveJson",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "prenom",
                                    "date_de_naissance",
                                    "genre",
                                    "telephone",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "prenom": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "date_de_naissance": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-01-01"
                                    },
                                    "genre": {
                                        "type": "string",
                                        "example": "M"
                                    },
                                    "profile": {
                                        "type": "string",
                                        "example": "ELEVE"
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "example": "75000000"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john.doe@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "@test@password#2024"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Votre compte a été crée avec succès"
                                        },
                                        "data": {
                                            "properties": {
                                                "nom": {
                                                    "type": "string",
                                                    "example": "Doe"
                                                },
                                                "prenom": {
                                                    "type": "string",
                                                    "example": "John"
                                                },
                                                "date_de_naissance": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "1990-01-01"
                                                },
                                                "genre": {
                                                    "type": "string",
                                                    "example": "M"
                                                },
                                                "profile": {
                                                    "type": "string",
                                                    "example": "ELEVE"
                                                },
                                                "telephone": {
                                                    "type": "string",
                                                    "example": "75000000"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "format": "email",
                                                    "example": "john.doe@example.com"
                                                },
                                                "password": {
                                                    "type": "string",
                                                    "example": "@test@password#2024"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Les données fournies ne sont pas valides."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/verify-otp": {
            "post": {
                "tags": [
                    "Autentification"
                ],
                "summary": "Verification du code otp",
                "description": "Validation de l'adresse mail à travers le code OTP",
                "operationId": "7ffdcc2529f1d1bbbbac715c91b1c809",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "john.doe@example.com ou 75000000"
                                    },
                                    "otp": {
                                        "type": "string",
                                        "example": "0000"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Code OTP vérifié avec succès"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Les données fournies ne sont pas valides."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}